fix double compare problem

+++ src/compute/MTCompute/computeMt-openMP.cpp
@@ -49,7 +49,8 @@ vector2D LSH::computeCollision_openmp(vector2D hMatrixN, vector2D hMatrixQ){
     for (int q = 0; q <Q ; ++q) {
         for (int n = 0; n <N ; ++n){
             for (int hash_id = 0; hash_id < L; ++hash_id) {
-                if (hMatrixN[n][hash_id] == hMatrixQ[q][hash_id])
+                //cast to int for comparision, reduce double compare risk
+                if ((int)hMatrixN[n][hash_id] == (int)hMatrixQ[q][hash_id])
                     collisionMatrix[q][n]++;
             }
         }
